requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
python-dotenv==1.0.0
PyJWT==2.8.0
passlib==1.7.4
bcrypt